tuidan

+++ GSS3.0-IRS/gss3.0-product/gss-product-unp/gss-product-unp-api/src/main/java/com/tempus/gss/product/unp/api/entity/UnpSaleRefundItem.java
@@ -24,7 +24,7 @@ public class UnpSaleRefundItem implements Serializable {
     /**
      * 产品类型
      */
-    private Integer unpType;
+    private Long unpType;
     /**
      * 数量
      */
@@ -62,11 +62,11 @@ public class UnpSaleRefundItem implements Serializable {
         this.saleRefundOrderNo = saleRefundOrderNo;
     }
     
-    public Integer getUnpType() {
+    public Long getUnpType() {
         return unpType;
     }
     
-    public void setUnpType(Integer unpType) {
+    public void setUnpType(Long unpType) {
         this.unpType = unpType;
     }
